# docs/接口文档/微友App-openapi.yaml
openapi: 3.0.3
info:
  title: 微友 App Backend API Draft
  version: 0.1.0-draft
  description: |
    基于当前后端控制器、需求文档与已生成接口文档整理的 OpenAPI 3.0 草稿。
    
    说明：
    - HTTP Base URL: `http://localhost:8080/api`
    - WebSocket URL: `ws://localhost:8081/ws/chat`
    - WebSocket 协议详情见 `docs/接口文档/微友App-后端接口文档.md`
    - 当前部分接口仍为 demo/stub/no-op 实现，字段结构可用于联调，业务结果未必代表最终生产行为
servers:
  - url: http://localhost:8080/api
    description: Local development server
security:
  - bearerAuth: []
...
paths:
  /auth/login/password:
    post:
      tags:
        - Auth
      summary: 密码登录
      description: 匿名接口，标准密码登录。
      security: []
      operationId: authLoginPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginPasswordRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginTokenData'
  /auth/login/sms:
    post:
      tags:
        - Auth
      summary: 短信验证码登录
      description: 匿名接口。当前 smsCode 未真正校验，手机号不存在时会自动注册。
      security: []
      operationId: authLoginSms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginSmsRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginTokenData'
  /auth/register:
    post:
      tags:
        - Auth
      summary: 注册
      description: 匿名接口。当前 smsCode 未真正校验；若手机号已存在则直接返回原账号登录态。
      security: []
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginTokenData'
  /auth/sms/send:
    post:
      tags:
        - Auth
      summary: 发送短信验证码
      description: 匿名接口。当前为 no-op 占位实现。
      security: []
      operationId: authSendSms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsSendRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /auth/token/refresh:
    post:
      tags:
        - Auth
      summary: 刷新 token
      description: 匿名接口。仅接受 refresh token。
      security: []
      operationId: authRefreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginTokenData'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: 退出登录
      description: 按 deviceId 将目标设备下线。
      operationId: authLogout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceActionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /auth/device/offline:
    post:
      tags:
        - Auth
      summary: 下线指定设备
      description: 当前未校验设备归属。
      operationId: authDeviceOffline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceActionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /collection/list:
    get:
      tags:
        - Collection
      summary: 查询收藏列表
      operationId: collectionList
      parameters:
        - in: query
          name: type
          required: false
          schema:
            type: string
        - in: query
          name: keyword
          required: false
          schema:
            type: string
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            format: int32
            default: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageCollectionItemData'
  /collection/create:
    post:
      tags:
        - Collection
      summary: 创建收藏
      operationId: collectionCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCollectionItem'
  /collection/delete:
    post:
      tags:
        - Collection
      summary: 删除收藏
      operationId: collectionDelete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionDeleteRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCollectionItemList'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        traceId:
          type: string
          example: 3f4b9d6e8a2c4f1b
...
    PageDataBase:
      type: object
      required:
        - pageNo
        - pageSize
        - total
        - hasMore
      properties:
        pageNo:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        hasMore:
          type: boolean
        nextCursor:
          type: string
          nullable: true
...
    ApiResponseVoid:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              nullable: true
    LoginPasswordRequest:
      type: object
      required:
        - mobile
        - password
        - deviceId
      properties:
        mobile:
          type: string
        password:
          type: string
        deviceId:
          type: string
    LoginSmsRequest:
      type: object
      required:
        - mobile
        - smsCode
        - deviceId
      properties:
        mobile:
          type: string
        smsCode:
          type: string
        deviceId:
          type: string
    RegisterRequest:
      type: object
      required:
        - mobile
        - smsCode
        - password
      properties:
        mobile:
          type: string
        smsCode:
          type: string
        password:
          type: string
        deviceId:
          type: string
          nullable: true
    SmsSendRequest:
      type: object
      required:
        - mobile
        - scene
      properties:
        mobile:
          type: string
        scene:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    DeviceActionRequest:
      type: object
      required:
        - deviceId
      properties:
        deviceId:
          type: string
    LoginTokenData:
      type: object
      required:
        - userId
        - accessToken
        - refreshToken
        - expireIn
        - needDeviceVerify
      properties:
        userId:
          type: integer
          format: int64
        accessToken:
          type: string
        refreshToken:
          type: string
        expireIn:
          type: integer
          format: int32
        needDeviceVerify:
          type: boolean
...
    CollectionItem:
      type: object
      properties:
        collectionId:
          type: integer
          format: int64
        type:
          type: string
        title:
          type: string
        cover:
          type: string
          nullable: true
        summary:
          type: string
        createdAt:
          type: string
    CollectionCreateRequest:
      type: object
      required:
        - type
        - title
      properties:
        type:
          type: string
        title:
          type: string
        cover:
          type: string
        summary:
          type: string
    CollectionDeleteRequest:
      type: object
      required:
        - collectionId
      properties:
        collectionId:
          type: integer
          format: int64
...
    PageCollectionItemData:
      allOf:
        - $ref: '#/components/schemas/PageDataBase'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/CollectionItem'
    ApiResponseLoginTokenData:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoginTokenData'
...
    ApiResponsePageCollectionItemData:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageCollectionItemData'
    ApiResponseCollectionItem:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CollectionItem'
    ApiResponseCollectionItemList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/CollectionItem'
...